Make the merge tool handle "UA" conflicts.
This conflict type happens when upstream renames a file we deleted for
licensing reasons. Allow the file to be re-added in case it was renamed
to a path we don't delete; it will be deleted again in the next step if
it's still in a directory we delete.

Also, simplify the existing conflict resolution code; findall returns a
string if there's only one match group, but a tuple if there's more than
one.

Change-Id: I3b55b6549b746e9e530977d99ad541e0eac096e0