Merge scripts: merge from LKGR by default.
Fetch the last known good revision from the Chromium waterfall and use
it as the default revision to merge to. Fix the svn->git revision
conversion to use the earliest match, not the latest, as this is almost
always correct.

Also, rename Snapshot without an underscore as we're going to reuse it.

Change-Id: Idd0c9db40ca6b7212c6bbcca6046bf7b09ff3259