Merge scripts: support pushing via a temporary branch.
Push to a temporary branch first in order that the server has already
got a copy of all our objects when we push to the real branch, in an
effort to reduce skew between repositories. Move Push into a seperate
function also, so that it can be called from other scripts without
needing to duplicate the branch configuration.

Change-Id: I8eec01db5d0e1439e391b185b7250bbc23c3e3e3