Merge scripts: allow automation to be smarter.
1) Use a branch with the revision number in the name, so that if
something goes wrong we can try again at a later revision without
overwriting the previous branch, in case someone wants to look at it.

2) Split out the code to retrieve LKGR into a separate function that can
be reused, so we don't have to try and do the merge before discovering
there is nothing new.

3) Introduce a new exception, TemporaryMergeError, which specifically
identifies problems that may be fixable by just retrying the merge with
a later revision. Use git reset --hard to clean up broken states when we
are going to raise this.

Change-Id: Ib32a4c211447abeb3daefa5fcc61360e13f704fc