Fake an onNewPicture callback after onPageFinished.
The new upstream rendering path altered the timing on onNewPicture
callbacks such that some of the CTS tests became flaky, timing out
waiting for pages to load. Re-introduce a hack in the glue layer that
will fake an onNewPicture callback at the time the CTS tests expect
to deflake them.

Bug: 8208948
Change-Id: I21f20880099dd3d5f06079ba8662981764991455