Enable jarjar for webviewchromium.
When building webviewchromium.jar, use jarjar to rename classes in
accordance with the rules file checked into the Chromium tree. The same
rules must be used on both sides as the JNI generator needs to also
rename its class references when building native code. The proguard
rules must also reference the renamed classes, as jarjar runs before
proguard - we keep both sets of names in the proguard config so that
jarjar can be disabled if needed without disabling proguard.

Change-Id: Ib5675ead752c5c3cd62c9ae7bbf84106f4454288