Enable proguard.
CTS tests appear to pass at the expected rate with proguard enabled, so
turn it on for now so that we can get some thorough testing. Revert this
if it appears to be causing problems and you can't work out what to add
to the flag file to fix it.

Change-Id: I15e0d0d5b31861ddd0eef4bf21163df2e0d9a7fd