Improve proguard configuration.
1) We should explicitly keep the native methods in
com.android.webview.chromium as well, since the plat_support library
will crash if they are removed. This wasn't actually broken as they were
references anyway, but being explicit is better.

2) Using a single "keep" statement for the org.chromium JNI interface
causes all classes in that package to be kept even if they don't have
JNI members. Instead, we should have a separate "keepclasseswithmembers"
for each of the kinds of JNI interface to keep only classes that have
at least one of those methods. This allows a few more classes to be
removed entirely (along with their constructors) and prunes a small
amount of extra code as a result.

Change-Id: I7ce4dd7bdca36af49a306838bb567317e5314f4f