Add tool to check that we jarjar'ed everything.
Add a tool to verify that after jarjar has run, all the classes
remaining are in whitelisted packages. This will ensure that we don't
accidentally introduce a conflicting class definition into the system
classpath by failing to rename a new dependency. Make the built module
depend on this check step so that the build will fail if the check
doesn't pass.

Bug: 8030113
Change-Id: I92fc75eeee644a3eb05a13799ae398be20c286b3