Fix handling of non-data: base-URL loads.
When faced with a non-data:scheme base URL, the classic WebView
would simply dump the data string into the WebView (using the passed
mime type) without performing the usual URL loading steps such as
decoding URL encoded entities.

Fix the WebViewChromium implementation to also have this behavior -
prior to this CL we would load all data (regardless of the base URL)
passed to loadDataWithBaseUrl as a data: URL. This meant that inside
the Chromium stack we'd treat these loads as regular URL loads and
perform the URL entity decode. Some apps relied on this decode not
happening so this is primarily a backwards compatibility quirk.

Bug: 9894754

Change-Id: I942634a09dbcf6db799a1daf64ba334757b1fa79