Merge scripts: automate more of merge_to_master.
Allow merge_to_master.py to be given the path to a repo.prop file (from
the build system). This will be used to detect the revision of chromium
to merge automatically, and will also be used to create and upload a
merge CL for review for frameworks/webview. Also, update all the remotes
explictly when running the script to remove the need to sync both master
and master-chromium first; you can now just run the script on a
master-chromium checkout.

Change-Id: Ie812bac3f23864dde64ceebedb443e7afbe0571f