Ferrying setBackgroundColor to the UI thread
It is just a void setter so trivial. Marking getViewDelegate as
threadsafe since it is harder to move that to the UI thread, and
it is safe (just return this).

Finally throwing the thread violation exception on the UI thread
so it will (most likely) kill the app. Creating the exception on
the violating thread so we get the right stack trace.

BUG=10468752

Change-Id: If151594517406952342a64c89caddac334aa255a