Wire up stylable attributes for <select> dialogs.
This complements crrev.com/23523025 and overrides the layouts used
for rendering the <select> dialogs using the same ones used by classic
WebView. Clank's default layouts, in fact, can be (graphically) broken
by apps if they override, for instance as happened in b/10549550, the
text color for their alert dialogs and show the WebView in an alert dialog.

Change-Id: I3e7cb1a5422a5457bfe69906f43a43b2eb40a8e0