Add instrumentation-based WebViewFlingTest.
Previous test used the legacy shell-based uiautomator and
janktesthelper. The shell-based libraries relied on some hacks to work
around the fact that they did not have access to a Context object, but
these hacks were prone to being flaky. The new versions of the libraries
rely on Instrumentation and avoid these hacks by using a Context instead.

Bug: 13471491
Change-Id: I6faff306e475bed38ccaca377f099fa40334b77c