Merge scripts: run the libaddressinput hook.
The libaddressinput hook from DEPS needs to be run to make our build
work; copy it to the merge script for now as we still haven't got a
better way to deal with hooks. Remove the old hook that downloaded gn
from google storage as this is not necessary at present and is unlikely
to be needed in future either.

Change-Id: I04facd61fba554fcfc0093ef8bc907474d6c1d07