CHerry pick Include pak files in the WebView apk.
This change makes the webviewchromium.apk include the Chromium
native resource .pak files. The main reason for the complexity
of the change is the fact that the .pak files are generated by
the build system and therefore need to be copied to the target
folder before the AAPT invocation.

Original BUG=11505352

Original Change-Id: I42029f4bc11684962f01a7746fa0d34589962838

Bug: 16723226
Change-Id: Ie01a26e79f469b74b7c841ff8e4057b50d2308d8