Generate symlinks to libwebviewchromium.so in /system/priv-app
This change declares libwebviewchromium to be treated as a JNI lib
for the webviewchromium.apk. By doing this, the build system will
create symlinks from /system/app-priv/webviewchromium/lib/...
(which is the path returned by the package manager) to the actual
libraries in /system/lib*.
Furthermore this change also declares the apk to be a multilib
target, so that both libraries get symlinked on multi-arch devices.

BUG:16403706
Change-Id: Ib02191eddf8262b2b23c3c32d07e77b33759285d