Don't declare WebView as a library.
We don't need to declare WebView as a library as we add its resources to
the asset manager dynamically rather than through a <uses-library> tag
in an application manifest. This also removes the need for it to be a
privileged application.

Bug: 11505352
Change-Id: Ief2f7a72a1068f51edf8842bd1d1c77dad7b23ca
(cherry picked from commit f3f92fd925e1a70e93e07850b8d5a8d0b5328e08)