Don't declare WebView as a library.
We don't need to declare WebView as a library as we add its resources to
the asset manager dynamically rather than through a <uses-library> tag
in an application manifest. This also removes the need for it to be a
privileged application.

Bug: 11505352
Change-Id: I504af9dc120739368169653d71072251b907a41f
(cherry picked from commit bd8a80488b151ceee823eb8f06419a1991d53b8f)