Set target and min sdk level for webview package
Bug: 16902981

Since current L level is not yet set to 21, we need to set
our target and min sdk level to 20 to be able to upload our
package to PlayStore. We will update the level to 21
once the new level is set.

Change-Id: I31125c893e714d9eb41dc2b7d16bce700f056d41
(cherry picked from commit 6beb32a5aa03941431169ce7581273611f0c6862)