Delete profile data if the WebView is downgraded.
If the WebView package has been downgraded since the previous time we
were used in this application, it may not be safe to load data from the
old profile as Chromium doesn't support this. Detect the downgrade by
saving the version code of the package in a shared preference and
deleting the profile files if it's lower.

Bug: 16724751
Change-Id: I832dccadee1f27f8bf480a21200be584949755f0