Update the merge_to_master.py script for Git.
Changes the merge_to_master.py script to deal with the Chromium's
switch to Git.
Furhermore, this change introduces pedantic checks when merging a given
master-chromium snapshot (using a repo.prop file), verifying that the
projects SHAs in the archive match the SHAs in the repo.prop.
Shouldn't this happen, it analyzes the history of the mismatching
projects and prints out a detailed error message. In case of a mismatch,
the merge can be forced using the --force option.

Change-Id: I8909d064709921220afbed1cc55cbcc7c48ececa