Fix resource rewriting in apps with other shared libs.
The resource rewriter was inadvertantly rewriting the resources for the
first shared library it found in the current application whether that
was the WebView or not due to a copy and paste oversight. Actually check
the name of the package before rewriting its resources. This is the name
of the package as the R class sees it, which is always
com.android.webview even when the manifest package name has been
changed.

Also remove unused imports.

BUG=17539974

Change-Id: I6c35451748ee9c32011b9a8b885f7c9b0b2e915d