Don't build the platform libraries in a tapas build.
When building the WebView APK with tapas, don't depend on
libwebviewchromium_loader or libwebviewchromium_plat_support as these
are part of the platform, not part of the APK, and there's no reason to
build them in the tapas build.

Also remove libwebviewchromium from the list entirely as it's already
listed in LOCAL_JNI_SHARED_LIBRARIES and thus already gets built, and
remove an obsolete comment.

Change-Id: I2b68a0537a4c36c627e933b6aeec571b5f665203